Module Auto_format

Shared place to implement common fmt related commands, with the ability to control that the AST doesn't change, and reaches a fix point.

Also contains the necessary utils to generate dune files that can extend the dune fmt rule to auto-format files with custom syntax.

module type T = sig ... end

module type T_pp = sig ... end

module Config : sig ... end

val fmt_cmd : (module Config.S) -> (module T with type t = 'a) -> (module Parsing_utils.S with type t = 'a) -> (module T_pp with type t = 'a) -> Command.t

val find_files_in_cwd_by_extensions : cwd:_ Eio.Path.t -> extensions:Base.string Base.list -> Base.string Base.list

Find all the files in the current directory that have one of the supplied extensions.

val allow_changes : Base.bool Base.Lazy.t

The pretty-printer may sometimes be used to perform some automatic refactoring on the files it formats. This is only possible when the environment variable AUTO_FORMAT_ALLOW_CHANGES=true.