Module type Config.S

val language_id : Base.string

A canonical name for the language to be used in the command manual and error messages. Prefer a name with lowercase and dashes, such as "my-custom-language".

val extensions : Base.string Base.list

The list of file extensions that are handled by your parser / printer pair. This may be several depending on the language (e.g. ocaml has .ml and .mli files.). Extensions are expected to be given with their leading dot included. Example: [".custom"].